feat: Missing track codes added.

+++ TrackLibrary.Base/Straight.cs
@@ -4,5 +4,5 @@ public class Straight : Segment
 {
     public Guid? PortA { get; set; }
     public Guid? PortB { get; set; }
-    public uint LengthInMm { get; set; }
+    public double LengthInMm { get; set; }
 }
