+++ TrackLibrary.PikoA/DKW.cs
@@ -9,5 +9,5 @@ public class DKW : SwitchDoubleCrossover
 {
     public uint LengthInMm { get; set; } = 239;
     public double ArcInDegree { get; set; } = 15;
-    public int RadiusInMm { get; set; } = 908;
+    public double RadiusInMm { get; set; } = 908;
 }
